requirements.txt
# conda 25.7.0 --> environment with python 3.12.11 and following packages:

# Flask and Web Dependencies
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3

# HTTP and Networking
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
urllib3==2.5.0

# Machine Learning and Data Processing
faiss-cpu==1.12.0
filelock==3.19.1
hf-xet==1.1.10
huggingface-hub==0.35.0
joblib==1.5.2
llvmlite==0.45.1
numba==0.62.1
numpy==2.3.3
pandas==2.3.2
pynndescent==0.5.13
regex==2025.9.18
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
umap-learn==0.5.9.post2

# Data Visualization
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.0
kiwisolver==1.4.9
matplotlib==3.10.6
pyparsing==3.2.5
seaborn==0.13.2

# Utilities
fsspec==2025.9.0
mpmath==1.3.0
networkx==3.5
pillow==11.3.0
python-dateutil==2.8.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.16.0
sympy==1.14.0
threadpoolctl==3.6.0
typing_extensions==4.11.0
tzdata==2025.2
wheel==0.45.1